+++ README.md
@@ -1,1 +1,10 @@
 # diploma-marvel-app
+
+1. RandomChar element done
+2. CharList element done
+3. CharInfo element done
+4. CharLoading done
+5. Banner element done
+6. ComicsList element done
+7. Router done
+
