+++ marvel/src/App.js
@@ -1,36 +1,26 @@
-import {useState} from 'react';
+
+import {BrowserRouter as Router, Routes, Route} from 'react-router-dom';
 import Header from './components/Header';
-import RandomChar from './components/RandomChar';
-import CharList from './components/CharList';
-import CharInfo from './components/CharInfo';
-import Banner from './components/Banner';
-import ComicsList from './components/ComicsList';
+import MainPage from './components/Pages/MainPage';
+import ComicsPage from './components/Pages/ComicsPage';
+import Page404 from './components/Pages/Page404';
 import './styles/variables.scss';
 
-import decoration from './resources/vision.png';
-
 const App = () => {
-  
-  const [selectedChar, setChar] = useState(null);
-
-  const onCharSelected = (id) => {
-    setChar(id);
-  }
-
+ 
   return (
-    <div className="App">
-      <Header/>
-      <main>
-        <RandomChar/>
-        <div className="char__content">
-          <CharList onCharSelected={onCharSelected}/>
-          <CharInfo charId={selectedChar}/>
-        </div>
-        <img className="bg-decoration" src={decoration} alt="vision"/>
-        <Banner/>
-        <ComicsList/>
-      </main>
+    <Router>
+      <div className="App">
+        <Header/>
+        <main>
+          <Routes>
+            <Route path="/comics" element={<ComicsPage/>}></Route>
+            <Route path="/" element={<MainPage/>}></Route>
+            <Route path="*" element={<Page404/>}></Route>
+          </Routes>
+        </main>
     </div>
+    </Router>
   );
 }
