+++ marvel/src/components/Header/Header.jsx
@@ -1,19 +1,28 @@
+import {Link, NavLink} from 'react-router-dom';
+
 import './Header.scss';
+import '../../styles/variables.scss';
 
 const Header = () => {
 
     return (
         <header className="app__header">
             <h1 className="app__title">
-                <a href="#">
+                <Link to="/">
                     <span>Marvel</span> information portal
-                </a>
+                </Link>
             </h1>
             <nav className="app__menu">
                 <ul>
-                    <li><a href="#">Characters</a></li>
+                    <li><NavLink 
+                        to="/"
+                        style={ ({ isActive }) => ({ color: isActive ? "#9F0013" : "#232222" }) }
+                        >Characters</NavLink></li>
                     /
-                    <li><a href="#">Comics</a></li>
+                    <li><NavLink 
+                        to="/comics"
+                        style={ ({ isActive }) => ({ color: isActive ? "#9F0013" : "#232222" }) }
+                        >Comics</NavLink></li>
                 </ul>
             </nav>
         </header>
