+++ marvel/src/components/Pages/ComicsPage/ComicsPage.jsx
@@ -1,0 +1,14 @@
+import Banner from '../..//Banner';
+import ComicsList from '../../ComicsList';
+
+const ComicsPage = () => {
+
+    return (
+        <>
+        <Banner/>
+        <ComicsList/>
+      </>
+    )
+}
+
+export default ComicsPage;
