+++ marvel/src/components/Pages/ComicsPage/index.js
@@ -1,0 +1,3 @@
+import ComicsPage from './ComicsPage';
+
+export default ComicsPage;
