+++ marvel/src/components/Pages/Page404/index.js
@@ -1,0 +1,3 @@
+import Page404 from './Page404';
+
+export default Page404;
