+++ marvel/src/styles/variables.scss
@@ -1,3 +1,3 @@
 $main-color: #9F0013;
 $dark: #232222;
-$grey: #5C5C5C;
+$grey: #797979;
